% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonaccis.R
\name{fibonaccis}
\alias{fibonaccis}
\title{Function to make integer vector of Fibonacci numbers}
\usage{
fibonaccis(t = 10)
}
\arguments{
\item{t}{a positive integer}
}
\description{
This function will make an integer vector
of with values from the Fibonacci sequence (up to the
first 25 values).
}
\examples{
fibonaccis()
}
\keyword{fibonacci}
