% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{is.prime}
\alias{is.prime}
\title{Test for Primality}
\usage{
is.prime(p)
}
\arguments{
\item{p}{a positive integer}
}
\description{
This function tests if a positive integer is prime or not.
}
\examples{
is_prime()
}
