% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleornot.R
\name{triangleornot}
\alias{triangleornot}
\title{Triange inequality tester}
\usage{
triangleornot(s1, s2, s3)
}
\arguments{
\item{s1}{a positive real number}

\item{s2}{a positive real number}

\item{s3}{a positive real number}
}
\description{
This function tests if 3 positive real numbers fulfil
the triangle triangle inequality. Result is printed.
}
\examples{
triangleornot()
}
\keyword{triangle}
